Move lifecycle stuff to FastArduino.
