+++ cores/fastarduino/lifecycle.cpp
